Christmas Commit
A lot of work on the main file, removed all static calls (was tossing
static member errors on test).  Command structure for static commands
set, user defined commands separated from the main list.  Much of the
SQL work is now done and responders are now in for system messages.  All
that is technically needed is the rest of the read checks and the user
command handlers as well as the callbacks for hard-coded commands.