Twitch Module
Completed the Twitch module with the current commands.  Entire
permissions system restructured to allow the live enabling and disabling
of any and all commands at any time, as well as modules outside of user
and core (protected).  All settings are saved and will be loaded upon
the reshell of the bot.  All environment data (Except the password when
connecting to Twitch) is now completely in construction and is REQUIRED
to be there.  All command registration is now in init and MUST be there
for any command permissions to work (Current setup will unset any and
all commands registered in construction phase).  Post init phase added
which overlays the channel permissions and settings over commands and is
not able to be interacted with.  2 new tables added for the Twitch
module.  Database module has new and standardized constructors for
queries and all SQL calls should be built with the constructors.
Startup order changed.  All SQL calls shifted to use new constructors.
New permission layer Turbo added.