Version 1.0 update
New filestructure in place to allow modules and dependencies to be
loaded without knowing what they are called or needing code to be added
to the bot.  Security to restrict what a module can do will be added in
later on.  New logger allows modules to tag their own output to make
finding their output easier and stack traces on most errors (Some fatal
errors will not have stack traces depending on how they force the script
to exit).  Module structure now standardized and all modules have access
to tick() and _read().  Compatibility for inspircd added.  New command
!topic added for standard IRC channels.  Reminders fixed to allow
reminders to properly cycle again and to allow them to be added and
removed from the stacks properly.