Moderation functionality update
Moderation module is now fully capable of moderating any channel, both
for standard IRC and for Twitch.tv.  The way it moderates a channel, as
in the number of kicks before a ban is customizable.  All filters are
able to bypass any permission layers and can be disabled on the fly like
modules and commands are.  Includes a word/phrase filter to filter out
specific words and phrases.  Currently uses hard-coded feedback into
chat, but there is some base framework in place to allow me to add
custom responses for each session ID.  IRC library made more strict on
checks to stop false positives when a user JOIN's a channel with a
hostname or nick containing part of a service message (IE modest would
trip the code to detect a MODE message).  Starting work in place for
command blacklisting to stop modules from running commands in either
core or other modules based on weather they are blacklisted or not.
Will be an addition to the command registry, but will be a new function
(Will work the exact same way as command registration does now, format
will be apparent when that update rolls out).  Reminders received fixes
to commands !editreminder, !reminders_addcommand, !reminders_delcommand
and !reminders_editcommand where the array was not being set properly
when attempting to update the information.  Reminders also received new
vars for an update where the module will stop processing messages if the
channel no longer has activity.  Reminders also no longer process every
2 minutes.  They now process as soon as the reminder is allowed to be
sent assuming the 2 minute timer has expired, meaning that a reminder
that is allowed to run every 3 minutes will run every 3 minutes instead
of every 4.