Update [09-05-2014]
Core Update to build 185.  Fixes JOIN detection and AUTH detection to be
both more reliable and more timed to the actual event.  Fixes modules
not ticking properly.  Fixes cases where no topic is sent properly to
the bot not populating the command storage for !topic.  Fixes command
delimeter settings not being loaded on startup.  Removes the database
object accessor since it is not needed.  Adds logging to pre-join
commands.  Fixes reconnect code NOT resetting the AUTH and JOIN state
vars.  Fixes all kick related code for detection and layer removal.
Removes initial mode detection.  Fixes permissions on runCommand().
runCommand() is now a public function as intended.  Fixes error
functions from the IRC library tossing errors when the resource was
released.  Adds code to properly split the topic message into the
provided parts.  Fixes the splitting for KICK messages.  Module channel:
Build 43.  Adds in the missing tick() function.  Izl (Example module)
moved into the new directory explained later in the startup changes.
Nickprotect: Build 9.  Fixes the module comment on the class.  Fixes
deleting nicks actually inserting the name again into DB as opposed to
the intended action.  Fixes detection of netsplit cases for QUITs.
Reminders: build 10.  Fixes reminders ticking when there are no
reminders to process.  Fixes !listrem not stating that there were no
reminders registered.  Startup (This is VERY important.  Most of the
above can be ignore, but this can NOT as it provides a peek into things
to come).  Fixes user aborts capable of stopping the bot from starting
properly.  Adds a new constant [BURNBOT_CORE_PATH] used to properly
detect modules and directories across all platforms.  Changes all module
loading checks and includes to use the new constant.  Changes how
modules are loaded.  Global modules are now placed in the base
directory.  Global modules are loaded on ALL instances of the bot, no
matter what network or channel.  Networks may now have a folder placed
in the module directory.  Any module within the network directory are
loaded for all channels within that network.  Inside of the network
folder a channel folder may exist.  And modules within the channel
folder are ONLY loaded for that channel.  This means that modules may
now be channel specific, making custom modules for channels more
feasible.  This system sets up some of the base framework as well for
external and modules in other languages.  More on that system will be
released as more of core supports that and the standards are documented.
The standards WILL be ported to the GIT for all to see.  External
modules will not have some access to core and will not work in the exact
same manner as modules that are loaded into core do.  Fixes
errorHandler() actually ignoring some error suppression.